Share one HeadWords type across the head-identity consumers
